pyproject.toml
[tool.poetry]
name = "python-100-days"
version = "0.1.0"
description = ""
authors = ["kdanielw <kdanielalvesw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
colorgram-py = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
